+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix MosaicRasterSource, GDALRasterSource and GeoTiffResampleRasterSource behavior [#3252](https://github.com/locationtech/geotrellis/pull/3252)
 - HttpRangeReader should live outside of the Spark package [#3254](https://github.com/locationtech/geotrellis/issues/3254)
 - HttpRangeReader moved to `geotrellis.util [#3256](https://github.com/locationtech/geotrellis/issues/3256)
+- Consistently construct GridExtents with `math.round` [#3248](https://github.com/locationtech/geotrellis/issues/3248)
 
 ## [3.3.0] - 2020-04-07